requirements.txt
Django==1.4.3
Jinja2==2.6
PIL==1.1.7
PyYAML==3.10
Pygments==1.6rc1
South==0.7.6
Sphinx==1.1.3
Whoosh==2.4.1
amqp==1.0.8
anyjson==0.3.3
billiard==2.7.3.21
boto==2.8.0
celery==3.0.15
coverage==3.6
distribute==0.6.34
dj-database-url==0.2.1
django-audiofield==0.5.4
django-celery==3.0.11
django-dajaxice==0.5.5
django-filer==0.9.3
django-haystack==1.2.7
django-mptt==0.5.5
django-openid-auth==0.4
django-polymorphic==0.3.1
django-reusableapps==0.1.1
django-social-auth==0.7.18
django-storages==1.1.8
django-uuidfield==0.4.0
django-videothumbs==0.91
docutils==0.10
easy-thumbnails==1.2
elementtree==1.2.7-20070827-preview
gunicorn==0.17.2
httplib2==0.7.7
kombu==2.5.6
oauth2==1.5.211
psycopg2==2.4.5
pyelasticsearch==0.3
pysolr==3.0.5
python-dateutil==1.5
python-openid==2.2.5
pytz==2012j
requests==1.1.0
simplejson==3.1.0
switch2bill-common==2.6.0
virtualenv==1.8.2
